console: Adapt gateway-data-form
